client side buildfix
