+++ deprecated/drawHeroes.js
