+++ deprecated/drawTeammates.js
