+++ public/js/generateCharts.js
@@ -1,4 +1,4 @@
-module.exports = function generateCharts(data) {
+window.generateCharts = function generateCharts(data) {
     var color_array = [];
     for (var key in constants.player_colors) {
         color_array.push(constants.player_colors[key]);
