+++ test/test.js
@@ -5,6 +5,8 @@ config.POSTGRES_URL = "postgres://postgres:postgres@localhost/template1";
 config.REDIS_URL = "redis://localhost:6379/1";
 config.SESSION_SECRET = "testsecretvalue";
 config.NODE_ENV = "test";
+config.ENABLE_MATCH_CACHES = 1;
+config.ENABLE_PLAYER_CACHES = 1;
 var async = require('async');
 var redis = require('../redis');
 var queue = require('../queue');
