+++ views/base_layout.jade
@@ -45,23 +45,36 @@ html
     include navbar
     block page_content
     include footer
+
+
+  //-
+  script(src="//code.jquery.com/jquery-2.1.4.min.js")
+  script(src="//cdn.rawgit.com/mbostock/d3/master/d3.min.js")
+  link(rel='stylesheet' href='//cdn.rawgit.com/brianreavis/selectize.js/v0.12.1/dist/css/selectize.bootstrap3.css')
+  script(src="//cdn.rawgit.com/brianreavis/selectize.js/v0.12.1/dist/js/standalone/selectize.js")
+  link(rel='stylesheet' href="//cdn.rawgit.com/wa0x6e/cal-heatmap/3.5.3/cal-heatmap.css")
+  script(src="//cdn.rawgit.com/wa0x6e/cal-heatmap/3.5.3/cal-heatmap.min.js")
+  link(rel='stylesheet' href='//cdn.rawgit.com/masayuki0812/c3/0.4.11-rc4/c3.min.css')
+  script(src="//cdn.rawgit.com/masayuki0812/c3/0.4.11-rc4/c3.min.js")
+  //link(rel='stylesheet' href="//cdn.rawgit.com/DataTables/DataTables/1.10.9/media/css/jquery.dataTables.min.css")
+  link(rel='stylesheet' href="//cdn.rawgit.com/DataTables/DataTables/1.10.9/media/css/dataTables.bootstrap.css")
+  script(src="//cdn.rawgit.com/DataTables/DataTables/1.10.9/media/js/jquery.dataTables.min.js")
+  script(src="//cdn.rawgit.com/DataTables/DataTables/1.10.9/media/js/dataTables.bootstrap.min.js")
+  link(rel='stylesheet' href="https://cdn.rawgit.com/yasp-dota/qTip2/2.2.2/dist/jquery.qtip.min.css")
+  script(src="https://cdn.rawgit.com/yasp-dota/qTip2/2.2.2/dist/jquery.qtip.min.js")
+  
+  script(src="/public/js/playerMatches.js")
+  script(src="/public/js/tooltips.js")
+  script(src="/public/js/formatHtml.js")
+  script(src="/public/js/buildMap.js")
+  script(src='/public/js/createHistogram.js')
+  script(src='/public/js/createCalHeatmap.js')
+  script(src='/public/js/ratingsChart.js')
+  script(src='/public/js/generateCharts.js')
+  script(src='/public/js/util.js')
+  
   //-our custom js bundle, dynamically inserts external library js/css
   script(src="/public/build/yasp.min.js")
-  
-  //-
-    script(src="//code.jquery.com/jquery-2.1.4.min.js")
-    link(rel='stylesheet' href='//cdn.rawgit.com/brianreavis/selectize.js/v0.12.1/dist/css/selectize.bootstrap3.css')
-    script(src="//cdn.rawgit.com/brianreavis/selectize.js/v0.12.1/dist/js/standalone/selectize.js")
-    //script(src="//cdn.rawgit.com/mbostock/d3/master/d3.min.js")
-    //script(src="//cdn.rawgit.com/wa0x6e/cal-heatmap/master/cal-heatmap.min.js")
-    link(rel='stylesheet' href='//cdn.rawgit.com/masayuki0812/c3/0.4.11-rc4/c3.min.css')
-    script(src="//cdn.rawgit.com/masayuki0812/c3/0.4.11-rc4/c3.min.js")
-    
-    link(rel='stylesheet' href="//cdn.rawgit.com/DataTables/DataTables/1.10.9/media/css/jquery.dataTables.min.css")
-    link(rel='stylesheet' href="//cdn.rawgit.com/DataTables/DataTables/1.10.9/media/css/dataTables.bootstrap.css")
-    script(src="//cdn.rawgit.com/DataTables/DataTables/1.10.9/media/js/jquery.dataTables.min.js")
-    script(src="//cdn.rawgit.com/DataTables/DataTables/1.10.9/media/js/dataTables.bootstrap.min.js")
-
 
   //stuff to include after page load, js
   block footer_assets
