+++ views/player/player_items.jade
@@ -29,7 +29,7 @@ block playercontent
                 else
                   =key
               td=player.aggData.purchase.counts[key]
-              td=player.aggData.item_usage.counts[key]
+              td=player.aggData.item_usage.counts[key] 
               td=player.aggData.item_uses.counts[key]
               td=player.aggData.item_uses.counts[key] ? (player.aggData.item_uses.counts[key]/player.aggData.item_usage.counts[key]).toFixed(2) : ""
               td=val/player.aggData.purchase.counts[key]
